+++ sources/RobertMunteanu/Vending machine/Product.cs
@@ -5,11 +5,11 @@ namespace VendingMachine
     public class Product
     {
         public string Name { get; private set; }
-        public double Price { get; set; }
+        public Decimal Price { get; set; }
         public Category Category{ get; private set; }
         public int Quantity { get; set; }
 
-        public Product(string Name, double Price, Category Category, int Quantity)
+        public Product(string Name, Decimal Price, Category Category, int Quantity)
         {
             this.Name = Name;
             this.Price = Price;
